addDependencies should be a PackageGraph instance method
